pyproject.toml
[project]
name = "ir2mqtt"
version = "1.0.0-beta.1"
description = "IR to MQTT bridge"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "paho-mqtt",
    "pydantic",
    "pydantic-settings",
    "python-multipart",
    "httpx",
    "sqlalchemy",
    "alembic",
    "greenlet",
    "aiosqlite",
    "pyserial",
    "pyserial-asyncio",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pip-tools",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "pytest-qt",
    "pytest-mock",
    "PyQt6",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "backend"}

[tool.ruff]
line-length = 180
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []